command refactoring
move gyv command to the project root
move create package into command package
fix defer lint errors
refactor errors output
